backend/requirements.txt
PyMySQL==0.9.2
bcrypt==3.1.2
Flask-Bcrypt==0.7.1
xlrd==1.1.0
Werkzeug==0.11.10
Flask-Cors==3.0.6
